Add score display
Add a big score display to the game scene. It automatically ticks up
by whatever the current "combo" is. For now I'm just setting it to be
the coolometercount in the gamescene update.

Also exposes an interface to modify the current combo, or to toggle
tweens which make it behave crazier.