Add overlays system
Added a flexible system for overlays to display on top of the game
while paused.

The `OverlayManager` object controls the opening and closing of the
overlays (which each extend the base `Overlay` class).

Need to do some more work to create the other overlays, add the
credits, figure out how to make game options work etc.