Hook up achievements
The achievements are now achievable (except the last one,
intentionally).

There are a couple of @NOTEs describing bugs in how this is
implemented, but I don't think it's worth chasing them, the happy path
works well enough.

Also cranked the score to increase by the cube of the combo meter to
really get those numbers up.